+++ pkg/events/redis/redis.go
@@ -18,6 +18,7 @@ package redis
 import (
 	"context"
 	"encoding/json"
+	"strings"
 	"sync"
 
 	"github.com/go-redis/redis/v8"
@@ -28,53 +29,82 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/log"
 	ttnredis "go.thethings.network/lorawan-stack/v3/pkg/redis"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
+	"go.thethings.network/lorawan-stack/v3/pkg/unique"
 )
 
 // NewPubSub creates a new PubSub that publishes and subscribes to Redis.
 func NewPubSub(ctx context.Context, taskStarter component.TaskStarter, conf ttnredis.Config) *PubSub {
 	ttnRedisClient := ttnredis.New(&conf)
-	eventChannel := ttnRedisClient.Key("events")
+	eventChannel := func(ctx context.Context, name string, ids *ttnpb.EntityIdentifiers) string {
+		if name == "" {
+			name = "*"
+		}
+		if ids == nil {
+			return ttnRedisClient.Key("events", "*", "*", name)
+		}
+		return ttnRedisClient.Key("events", ids.EntityType(), unique.ID(ctx, ids), name)
+	}
 	ctx = log.NewContextWithFields(ctx, log.Fields(
 		"namespace", "events/redis",
-		"channel", eventChannel,
 	))
 	ctx, cancel := context.WithCancel(ctx)
-	return &PubSub{
-		PubSub:       basic.NewPubSub(),
-		taskStarter:  taskStarter,
-		ctx:          ctx,
-		cancel:       cancel,
-		client:       ttnRedisClient.Client,
-		eventChannel: eventChannel,
+	ps := &PubSub{
+		PubSub:        basic.NewPubSub(),
+		ctx:           ctx,
+		cancel:        cancel,
+		client:        ttnRedisClient.Client,
+		eventChannel:  eventChannel,
+		subscriptions: make(map[string]int),
 	}
+	ps.sub = ps.client.Subscribe(ctx)
+	taskStarter.StartTask(&component.TaskConfig{
+		Context: ps.ctx,
+		ID:      "events_redis_subscribe",
+		Func:    ps.subscribeTask,
+		Restart: component.TaskRestartOnFailure,
+		Backoff: component.DefaultTaskBackoffConfig,
+	})
+	return ps
 }
 
 // PubSub with Redis backend.
 type PubSub struct {
-	events.PubSub
+	*basic.PubSub
 
-	taskStarter  component.TaskStarter
-	ctx          context.Context
-	cancel       context.CancelFunc
-	eventChannel string
-	client       *redis.Client
-	subOnce      sync.Once
+	eventChannel func(ctx context.Context, name string, ids *ttnpb.EntityIdentifiers) string
+
+	ctx    context.Context
+	cancel context.CancelFunc
+	client *redis.Client
+
+	subOnce       sync.Once
+	mu            sync.RWMutex
+	sub           *redis.PubSub
+	subscriptions map[string]int
 }
 
 var errChannelClosed = errors.DefineAborted("channel_closed", "channel closed")
 
 func (ps *PubSub) subscribeTask(ctx context.Context) error {
 	logger := log.FromContext(ctx)
-	sub := ps.client.Subscribe(ctx, ps.eventChannel)
-	logger.Info("Subscribed")
+
+	ps.mu.Lock()
+	ps.sub = ps.client.Subscribe(ctx)
+	patterns := make([]string, 0, len(ps.subscriptions))
+	for pattern := range ps.subscriptions {
+		patterns = append(patterns, pattern)
+	}
+	logger.WithField("patterns", patterns).Debug("Subscribe to Redis channels")
+	ps.sub.PSubscribe(ctx, patterns...)
+	ps.mu.Unlock()
+
 	defer func() {
-		if err := sub.Close(); err != nil {
+		if err := ps.sub.Close(); err != nil {
 			logger.WithError(err).Warn("Failed to close Redis subscription")
-		} else {
-			logger.Info("Unsubscribed")
 		}
 	}()
-	ch := sub.Channel()
+
+	ch := ps.sub.Channel()
 	for {
 		select {
 		case <-ctx.Done():
@@ -88,23 +118,88 @@ func (ps *PubSub) subscribeTask(ctx context.Context) error {
 				logger.WithError(err).Warn("Failed to unmarshal event from JSON")
 				continue
 			}
-			ps.PubSub.Publish(evt)
+			ps.PubSub.Publish(&patternEvent{Event: evt, pattern: msg.Pattern})
 		}
 	}
 }
 
+type patternEvent struct {
+	events.Event
+	pattern string
+}
+
+func (ps *PubSub) eventChannelPatterns(ctx context.Context, name string, ids []*ttnpb.EntityIdentifiers) []string {
+	if name == "" {
+		name = "*"
+	} else {
+		name = strings.Replace(name, "**", "*", -1)
+	}
+
+	if len(ids) == 0 {
+		ids = []*ttnpb.EntityIdentifiers{nil}
+	}
+
+	var patterns []string
+	for _, id := range ids {
+		patterns = append(patterns, ps.eventChannel(ctx, name, id))
+		if appID := id.GetApplicationIDs(); appID != nil {
+			pattern := ps.eventChannel(ctx, name, ttnpb.EndDeviceIdentifiers{
+				ApplicationIdentifiers: *appID,
+				DeviceID:               "*",
+			}.EntityIdentifiers())
+			patterns = append(patterns, pattern)
+		}
+	}
+
+	return patterns
+}
+
 // Subscribe implements the events.Subscriber interface.
 func (ps *PubSub) Subscribe(ctx context.Context, name string, ids []*ttnpb.EntityIdentifiers, hdl events.Handler) error {
-	ps.subOnce.Do(func() {
-		ps.taskStarter.StartTask(&component.TaskConfig{
-			Context: ps.ctx,
-			ID:      "events_redis_subscribe",
-			Func:    ps.subscribeTask,
-			Restart: component.TaskRestartOnFailure,
-			Backoff: component.DefaultTaskBackoffConfig,
-		})
-	})
-	return ps.PubSub.Subscribe(ctx, name, ids, hdl)
+	ps.mu.Lock()
+	defer ps.mu.Unlock()
+
+	sub := &subscription{
+		name:     name,
+		ids:      ids,
+		patterns: ps.eventChannelPatterns(ctx, name, ids),
+		hdl:      hdl,
+	}
+
+	ps.PubSub.AddSubscription(sub)
+
+	for _, pattern := range sub.patterns {
+		if ps.subscriptions[pattern] == 0 {
+			err := ps.sub.PSubscribe(ps.ctx, pattern)
+			if err != nil {
+				return err
+			}
+		}
+		ps.subscriptions[pattern]++
+	}
+
+	go func() {
+		<-ctx.Done()
+
+		ps.PubSub.RemoveSubscription(sub)
+
+		ps.mu.Lock()
+		defer ps.mu.Unlock()
+
+		for _, pattern := range sub.patterns {
+			ps.subscriptions[pattern]--
+			if ps.subscriptions[pattern] == 0 {
+				log.FromContext(ctx).WithField("pattern", pattern).Debug("Unsubscribe from Redis channels")
+				err := ps.sub.PUnsubscribe(ps.ctx, pattern)
+				if err != nil {
+					log.FromContext(ps.ctx).WithField("pattern", pattern).WithError(err).Warn("Could not unsubscribe")
+				}
+				delete(ps.subscriptions, pattern)
+			}
+		}
+	}()
+
+	return nil
 }
 
 // Close the Redis publisher.
@@ -129,7 +224,15 @@ func (ps *PubSub) Publish(evt events.Event) {
 		logger.WithError(err).Warn("Failed to marshal event to JSON")
 		return
 	}
-	if err := ps.client.Publish(ps.ctx, ps.eventChannel, b).Err(); err != nil {
-		logger.WithError(err).Warn("Failed to publish event")
+	if ids := evt.Identifiers(); len(ids) > 0 {
+		for _, id := range ids {
+			if err := ps.client.Publish(ps.ctx, ps.eventChannel(evt.Context(), evt.Name(), id), b).Err(); err != nil {
+				logger.WithError(err).Warn("Failed to publish event")
+			}
+		}
+	} else {
+		if err := ps.client.Publish(ps.ctx, ps.eventChannel(evt.Context(), evt.Name(), nil), b).Err(); err != nil {
+			logger.WithError(err).Warn("Failed to publish event")
+		}
 	}
 }